# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: University Faculty Work Portal API
  description: |
    API для системи планування, звітування та оцінювання роботи викладачів університету.
    
    **Типи користувачів:**
    - GUEST - незареєстрований користувач (публічний доступ)
    - TEACHER - викладач
    - DIRECTOR - завідувач кафедри/декан (має права викладача + додаткові)
    - ADMIN - адміністратор системи
    
    **Автентифікація:**
    - JWT tokens (Access: 30 хв, Refresh: 30 днів)
    - Hybrid strategy з Redis кешуванням
  version: 1.0.0
  contact:
    name: API Support
    email: api@stu.cn.ua
servers:
  - url: https://npp.stu.cn.ua/api/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        uuid:
          type: string
          format: uuid
        firstName:
          type: string
          example: "Олександр"
        lastName:
          type: string
          example: "Шевченко"
        patronymic:
          type: string
          example: "Миколайович"
        login:
          type: string
          example: "shevchenko.o.m"
        email:
          type: string
          format: email
          example: "shevchenko@stu.cn.ua"
        role:
          type: string
          enum: [TEACHER, DIRECTOR, ADMIN]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        details:
          $ref: '#/components/schemas/UserDetailsResponse'
        departments:
          type: array
          items:
            $ref: '#/components/schemas/UserDepartmentResponse'
...
    UserDetailsResponse:
      type: object
      properties:
        id:
          type: integer
        photo:
          type: string
          nullable: true
          description: URL фотографії
        degree:
          $ref: '#/components/schemas/DictionaryItemResponse'
        position:
          $ref: '#/components/schemas/DictionaryItemResponse'
        englishLevel:
          $ref: '#/components/schemas/DictionaryItemResponse'
        nasAffiliation:
          $ref: '#/components/schemas/DictionaryItemResponse'
        honors:
          $ref: '#/components/schemas/DictionaryItemResponse'
    UserDepartmentResponse:
      type: object
      properties:
        department:
          $ref: '#/components/schemas/DepartmentResponse'
        joint:
          type: integer
          nullable: true
          description: "Суміщення ставок"
        workload:
          type: number
          format: float
          nullable: true
          description: "Навантаження"
...
    ActivityResponse:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        year:
          type: integer
          example: 2024
        semester:
          type: integer
          minimum: 0
          maximum: 2
          description: "0 - весь рік, 1 - перший семестр, 2 - другий семестр"
        worksType:
          type: string
          enum: [SCIENCE, METHOD, ORG]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activityDetails:
          $ref: '#/components/schemas/ActivityDetailsResponse'
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
    ActivityDetailsResponse:
      type: object
      properties:
        id:
          type: integer
        numInOrd:
          type: integer
          nullable: true
          description: "Номер у наказі (необов'язково)"
        numInRegulations:
          type: integer
          description: "Номер у положенні (обов'язково)"
        coefficient:
          type: number
          format: float
          description: "Коефіцієнт для розрахунку рейтингу"
        normTime:
          type: number
          format: float
          description: "Нормований час для розрахунку годин"
        additionalDetails:
          type: object
          description: "JSON з додатковими полями для різних типів робіт"
          additionalProperties: true
        active:
          type: boolean
...
    CreateActivityRequest:
      type: object
      required:
        - description
        - year
        - semester
        - worksType
        - activityDetailsId
      properties:
        description:
          type: string
          minLength: 3
          maxLength: 1000
        year:
          type: integer
          minimum: 2020
          maximum: 2100
        semester:
          type: integer
          minimum: 0
          maximum: 2
        worksType:
          type: string
          enum: [SCIENCE, METHOD, ORG]
        activityDetailsId:
          type: integer
          description: "ID з довідника activity_details"
        coAuthorIds:
          type: array
          items:
            type: integer
          description: "ID співавторів (інших викладачів)"
    UpdateActivityRequest:
      type: object
      properties:
        description:
          type: string
        year:
          type: integer
        semester:
          type: integer
        activityDetailsId:
          type: integer
        coAuthorIds:
          type: array
          items:
            type: integer
    CreateActivityDetailsRequest:
      type: object
      required:
        - numInRegulations
        - coefficient
        - normTime
      properties:
        numInOrd:
          type: integer
          nullable: true
        numInRegulations:
          type: integer
        coefficient:
          type: number
          format: float
        normTime:
          type: number
          format: float
        additionalDetails:
          type: object
          additionalProperties: true

    # ==================== STRUCTURE SCHEMAS ====================
    DepartmentResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        depLogo:
          type: string
        depHead:
          type: string
        depPage:
          type: string
        active:
          type: boolean
        faculty:
          $ref: '#/components/schemas/FacultyResponse'
...
    FacultyResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        facLogo:
          type: string
        facHead:
          type: string
        facPage:
          type: string
        active:
          type: boolean
        institute:
          $ref: '#/components/schemas/InstituteResponse'
...
    InstituteResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        instLogo:
          type: string
        instHead:
          type: string
        instPage:
          type: string
        active:
          type: boolean
...
    DictionaryItemResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
          description: "Тільки для degrees"
        level:
          type: string
          description: "Тільки для english-levels"
        active:
          type: boolean
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
paths:
...
  /activities:
    get:
      tags:
        - Activities
      summary: Список робіт
      description: |
        Отримання списку робіт (фільтрація для власних або всіх).
        **Доступ:** Authenticated users
      operationId: getActivities
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          description: "Фільтр по користувачу (якщо не вказано - власні роботи)"
          schema:
            type: integer
        - name: worksType
          in: query
          description: "Фільтр по типу роботи"
          schema:
            type: string
            enum: [SCIENCE, METHOD, ORG]
        - name: year
          in: query
          description: "Фільтр по року"
          schema:
            type: integer
        - name: semester
          in: query
          description: "Фільтр по семестру"
          schema:
            type: integer
            minimum: 0
            maximum: 2
        - name: yearFrom
          in: query
          description: "Період з (рік)"
          schema:
            type: integer
        - name: yearTo
          in: query
          description: "Період до (рік)"
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Список робіт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Activities
      summary: Створення роботи
      description: |
        Створення нової роботи.
        **Доступ:** TEACHER, DIRECTOR
        **Обмеження:** read-only режим блокує створення
      operationId: createActivity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Роботу створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав або read-only режим активний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /activities/{id}:
    get:
      tags:
        - Activities
      summary: Отримання роботи за ID
      description: |
        Деталі роботи.
        **Доступ:** authenticated users
      operationId: getActivityById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Деталі роботи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Роботу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Activities
      summary: Оновлення роботи
      description: |
        Оновлення роботи.
        **Доступ:** тільки автор роботи (власник або співавтор)
        **Обмеження:** read-only режим блокує редагування
      operationId: updateActivity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityRequest'
      responses:
        '200':
          description: Роботу оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: |
            Недостатньо прав (тільки автор) або система в режимі read-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Роботу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Activities
      summary: Видалення роботи (soft delete)
      description: |
        Помітка роботи як неактивної.
        **Доступ:** тільки власник роботи (перший автор)
        **Обмеження:** read-only режим блокує видалення
      operationId: deleteActivity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Роботу видалено
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: |
            Недостатньо прав (тільки власник) або система в режимі read-only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Роботу не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== ACTIVITY DETAILS (DICTIONARIES) ====================
  /activity-details:
    get:
      tags:
        - Activities
      summary: Довідник типів робіт
      description: |
        Отримання списку всіх типів робіт з коефіцієнтами.
        **Доступ:** authenticated users (TEACHER, DIRECTOR, ADMIN)
      operationId: getActivityDetails
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 100
        - name: active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Список типів робіт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Activities
      summary: Створення типу роботи
      description: |
        Додавання нового типу роботи до довідника.
        **Доступ:** тільки ADMIN
      operationId: createActivityDetails
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityDetailsRequest'
      responses:
        '201':
          description: Тип роботи створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetailsResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /activity-details/{id}:
    get:
      tags:
        - Activities
      summary: Деталі типу роботи
      description: |
        **Доступ:** authenticated users
      operationId: getActivityDetailsById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Деталі типу роботи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetailsResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип роботи не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Activities
      summary: Оновлення типу роботи
      description: |
        Редагування типу роботи.
        **Доступ:** тільки ADMIN
      operationId: updateActivityDetails
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityDetailsRequest'
      responses:
        '200':
          description: Тип роботи оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetailsResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип роботи не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Activities
      summary: Видалення типу роботи
      description: |
        Soft delete типу роботи.
        **Доступ:** тільки ADMIN
      operationId: deleteActivityDetails
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Тип роботи видалено
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип роботи не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== DEPARTMENTS ====================
...